ai_icd_extraction/requirements.txt
# RAFgenAI dependencies
streamlit
pandas
langchain
langchain-core
langchain-google-genai
google-generativeai
pydantic
python-dotenv
pymupdf
pytesseract
pdf2image
Pillow
openpyxl
faiss-cpu
langchain-community
langchain-huggingface
sentence-transformers